+++ src/common/components/AllocationToolsMenuSection/AllocationToolsMenuSection.js
@@ -14,6 +14,7 @@ import {
 export const AllocationToolsMenuSection = ({
   disabled,
   onDownloadAllocationWorksheet,
+  onUploadAllocationWorksheet,
 }) => {
   const intl = useIntl();
 
@@ -36,6 +37,21 @@ export const AllocationToolsMenuSection = ({
           </Icon>
         </Button>
       </IfPermission>
+
+      <IfPermission perm="ui-finance.allocations.create">
+        <Button
+          buttonStyle="dropdownItem"
+          onClick={onUploadAllocationWorksheet}
+          disabled={disabled}
+        >
+          <Icon
+            size="small"
+            icon="report"
+          >
+            <FormattedMessage id="ui-finance.actions.uploadAllocationWorksheet" />
+          </Icon>
+        </Button>
+      </IfPermission>
     </MenuSection>
   );
 };
@@ -43,4 +59,5 @@ export const AllocationToolsMenuSection = ({
 AllocationToolsMenuSection.propTypes = {
   disabled: PropTypes.bool,
   onDownloadAllocationWorksheet: PropTypes.func.isRequired,
+  onUploadAllocationWorksheet: PropTypes.func.isRequired,
 };