+++ src/common/components/UploadAllocationWorksheetModal/index.js
@@ -1,0 +1,1 @@
+export { UploadAllocationWorksheetModal } from './UploadAllocationWorksheetModal';