+++ src/common/components/index.js
@@ -1,2 +1,3 @@
 export { AllocationToolsMenuSection } from './AllocationToolsMenuSection';
 export { DownloadAllocationWorksheetModal } from './DownloadAllocationWorksheetModal';
+export { UploadAllocationWorksheetModal } from './UploadAllocationWorksheetModal';