+++ src/components/BatchAllocations/CreateBatchAllocations/CreateBatchAllocations.js
@@ -1,0 +1,5 @@
+export const CreateBatchAllocations = () => {
+  return (
+    <div>create</div>
+  );
+};