+++ src/components/BatchAllocations/CreateBatchAllocations/index.js
@@ -1,0 +1,1 @@
+export { CreateBatchAllocations } from './CreateBatchAllocations';