+++ src/components/BatchAllocations/UploadBatchAllocations/UploadBatchAllocations.js
@@ -1,0 +1,5 @@
+export const UploadBatchAllocations = () => {
+  return (
+    <div>upload</div>
+  );
+};